+++ method-quartz-cg.h
@@ -121,10 +121,11 @@ int libgamma_quartz_cg_crtc_restore(libgamma_crtc_state_t *restrict);
  * @param   this    Instance of a data structure to fill with the information about the CRTC
  * @param   crtc    The state of the CRTC whose information should be read
  * @param   fields  OR:ed identifiers for the information about the CRTC that should be read
- * @return          Zero on success, -1 on error. On error refer to the error reports in `this`
+ * @return          Zero on success, -1 on error; on error refer to the error reports in `this`
  */
 LIBGAMMA_GCC_ONLY__(__attribute__((__nonnull__, __warn_unused_result__)))
-int libgamma_quartz_cg_get_crtc_information(libgamma_crtc_information_t *restrict, libgamma_crtc_state_t *restrict, int32_t);
+int libgamma_quartz_cg_get_crtc_information(libgamma_crtc_information_t *restrict,
+                                            libgamma_crtc_state_t *restrict, unsigned long long);
 
 /**
  * Get the current gamma ramps for a CRTC, `float` version
